metrics: promote v2 to main metrics library
Promotes metrics v2 to be the main metrics library and removes the
old implementations.
