+++ Cargo.toml
@@ -2,13 +2,11 @@
 members = [
     "atomics",
     "buffer",
-    "fastmetrics",
     "heatmap",
     "histogram",
     "logger",
     "metrics",
-    "metrics-v2",
-    "metrics-v2/derive",
+    "metrics/derive",
     "ratelimiter",
     "streamstats",
     "time",
