--- metrics-v2/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "rustcommon-metrics-v2"
-version = "0.1.0"
-edition = "2018"
-authors = ["Sean Lynch <seanl@twitter.com>"]
-license = "Apache-2.0"
-
-[features]
-heatmap = [ "rustcommon-heatmap", "rustcommon-atomics" ]
-
-[dependencies]
-linkme = "0.2.9"
-once_cell = "1.9.0"
-parking_lot = "0.11.2"
-
-rustcommon-metrics-derive = { path = "derive" }
-rustcommon-heatmap = { path = "../heatmap", optional = true }
-rustcommon-atomics = { path = "../atomics", optional = true }
-rustcommon-time = { path = "../time" }
