+++ metrics/src/counter.rs
