--- metrics/src/entry/mod.rs
@@ -1,113 +1,0 @@
-// Copyright 2020 Twitter, Inc.
-// Licensed under the Apache License, Version 2.0
-// http://www.apache.org/licenses/LICENSE-2.0
-
-use core::hash::Hash;
-use core::hash::Hasher;
-use core::marker::PhantomData;
-
-use crate::*;
-
-use rustcommon_atomics::Atomic;
-
-pub struct Entry<Value, Count>
-where
-    Value: crate::Value,
-    Count: crate::Count,
-    <Value as Atomic>::Primitive: Primitive,
-    <Count as Atomic>::Primitive: Primitive,
-    u64: From<<Value as Atomic>::Primitive> + From<<Count as Atomic>::Primitive>,
-{
-    name: String,
-    source: Source,
-    _value: PhantomData<Value>,
-    _count: PhantomData<Count>,
-}
-
-impl<Value, Count> Clone for Entry<Value, Count>
-where
-    Value: crate::Value,
-    Count: crate::Count,
-    <Value as Atomic>::Primitive: Primitive,
-    <Count as Atomic>::Primitive: Primitive,
-    u64: From<<Value as Atomic>::Primitive> + From<<Count as Atomic>::Primitive>,
-{
-    fn clone(&self) -> Self {
-        Self {
-            name: self.name.clone(),
-            source: self.source,
-            _value: self._value,
-            _count: self._count,
-        }
-    }
-}
-
-impl<Value, Count> Statistic<Value, Count> for Entry<Value, Count>
-where
-    Value: crate::Value,
-    Count: crate::Count,
-    <Value as Atomic>::Primitive: Primitive,
-    <Count as Atomic>::Primitive: Primitive,
-    u64: From<<Value as Atomic>::Primitive> + From<<Count as Atomic>::Primitive>,
-{
-    fn name(&self) -> &str {
-        &self.name
-    }
-
-    fn source(&self) -> Source {
-        self.source
-    }
-}
-
-impl<Value, Count> Hash for Entry<Value, Count>
-where
-    Value: crate::Value,
-    Count: crate::Count,
-    <Value as Atomic>::Primitive: Primitive,
-    <Count as Atomic>::Primitive: Primitive,
-    u64: From<<Value as Atomic>::Primitive> + From<<Count as Atomic>::Primitive>,
-{
-    fn hash<H: Hasher>(&self, state: &mut H) {
-        self.name.hash(state);
-    }
-}
-
-impl<Value, Count> From<&dyn Statistic<Value, Count>> for Entry<Value, Count>
-where
-    Value: crate::Value,
-    Count: crate::Count,
-    <Value as Atomic>::Primitive: Primitive,
-    <Count as Atomic>::Primitive: Primitive,
-    u64: From<<Value as Atomic>::Primitive> + From<<Count as Atomic>::Primitive>,
-{
-    fn from(statistic: &dyn Statistic<Value, Count>) -> Self {
-        Self {
-            name: statistic.name().to_string(),
-            source: statistic.source(),
-            _count: PhantomData,
-            _value: PhantomData,
-        }
-    }
-}
-impl<Value, Count> PartialEq for Entry<Value, Count>
-where
-    Value: crate::Value,
-    Count: crate::Count,
-    <Value as Atomic>::Primitive: Primitive,
-    <Count as Atomic>::Primitive: Primitive,
-    u64: From<<Value as Atomic>::Primitive> + From<<Count as Atomic>::Primitive>,
-{
-    fn eq(&self, other: &Self) -> bool {
-        self.name == other.name
-    }
-}
-
-impl<Value, Count> Eq for Entry<Value, Count>
-where
-    Value: crate::Value,
-    Count: crate::Count,
-    <Value as Atomic>::Primitive: Primitive,
-    <Count as Atomic>::Primitive: Primitive,
-    u64: From<<Value as Atomic>::Primitive> + From<<Count as Atomic>::Primitive>,
-{
-}
