+++ metrics/src/gauge.rs
