+++ metrics/src/lazy.rs
