--- metrics/src/traits/mod.rs
@@ -1,15 +1,0 @@
-// Copyright 2020 Twitter, Inc.
-// Licensed under the Apache License, Version 2.0
-// http://www.apache.org/licenses/LICENSE-2.0
-
-mod count;
-mod float_convert;
-mod primitive;
-mod statistic;
-mod value;
-
-pub use count::Count;
-pub use float_convert::FloatConvert;
-pub use primitive::Primitive;
-pub use statistic::Statistic;
-pub use value::Value;
