+++ metrics/tests/named_metric.rs
